Move toCoreFunctionMetadata to separate file
